--- a/buildSrc/src/main/groovy/openhouse.maven-publish.gradle
+++ b/buildSrc/src/main/groovy/openhouse.maven-publish.gradle
@@ -12,6 +12,10 @@ task javadocJar(type: Jar) {
   from javadoc.destinationDir
 }
 
+tasks.withType(GenerateModuleMetadata) {
+  enabled = false
+}
+
 [jar, sourcesJar, javadocJar].each { task ->
   task.from(rootDir) {
     include 'LICENSE'